+++ types.d.ts
@@ -2,5 +2,4 @@ export type { Command } from './types/command';
 export type { Option } from './types/option';
 export type { YaclilOptions } from './types/init';
 export type { ParsedCommand, ParsedOption, ParsedOptions, ParsedArgs, ParsedStandaloneOption } from './types/parsed';
-export type { Handler } from './types/handler';
 export type { Arg } from './types/arg';