+++ types/handler.d.ts
@@ -1,6 +1,0 @@
-import type { ParsedCommand } from "./parsed";
-
-/**
- * Handler function for commands
- */
-export type Handler = (event: ParsedCommand) => void;